requirements.txt
Django==5.1.4
djangorestframework==3.15.2
celery==5.4.0
drf-spectacular==0.28.0
flower==2.0.1
